Crate reader_for_microxml

reader for microXml

Things are changing fast. 2020-01-13 LucianoBestia ver.1.0.0.

There are many xml parsers/readers/tokenizers/lexers around, but I need something very small and simple for my simple html templates.
I found the existance of a standard (or W3C proposal) for MicroXml - dramatically simpler then the full Xml standard. Perfect for my use-case: I have small simple html files, that are microXml compatible.

microXml

MicroXML is a subset of XML. It is dramatically simpler.
https://www.xml.com/articles/2017/06/03/simplifying-xml-microxml/
https://dvcs.w3.org/hg/microxml/raw-file/tip/spec/microxml.html
MicroXML is actualy well-formed Xml.
In the data model of MicroXml there are no comments, CData, namespaces, declarations, processing instructions,...
An example of all can be done in a well-formed microXml:

<memo lang="en" date="2017-05-01">
    I <em>love</em> &#xB5;
    <!-- MICRO SIGN -->XML!<br />
    It's so clean &amp; simple.</memo>

MicroXml can be only in utf-8. I am lucky, because Rust Strings are internally utf-8 and are automatically checked for correctnes.
MicroXml should go throught normalization: CR & CRLF should be converted to LF, but I don't do that here.
MicroXml can contain Comments, but they are not microxml data, so I just skip them.
Whitespaces are completely preserved in Text Nodes. For me they are significant. Also newline and Tabs. This is different from full Xml whitespace processing.
All other whitespaces are ignored - they are insignificant.

reader

This ReaderForMicroXml obviously cannot read a complicated full XML.
This reader_for_microxml is used for small html fragments.
They must be well-formed microXml.
This fragments are meant for a html templating for dodrio.
Because of the small size of fragments, I can put all the text in memory in a string.
Only basic mal-formed incorrectness produce errors. I am not trying to return errors for all the possible mal-formed incorrectness in microXml.
The speed is not really important, but the size of the code is, because it will be used in WebAssembly. Every code is too big for Wasm!
The crate has NO dependencies, NO allocations.
Probably it could be also #![no_std], but I don't need that.

References

https://dvcs.w3.org/hg/microxml/raw-file/tip/spec/microxml.html
https://www.xml.com/articles/2017/06/03/simplifying-xml-microxml/
https://github.com/tafia/quick-xml
https://github.com/RazrFalcon/roxmltree

Structs

ReaderForMicroXml
struct Reader for MicroXml - the Class Rust has Structs + Traits, but for me it is just like Class/Object. Just without inheritance. All the fields are internal and not public. The only way to interact is throu methods.

Enums

Event
The reader_for_microxml returns events. The caller will manage this events. So they must be public. The string slices are reference to the original string with microxml text

TagState
internal enum: Tags are strings inside delimiters < and > like
or