Struct reader_for_microxml::ReaderForMicroXml

pub struct ReaderForMicroXml<'a> {
    input: &'a str,
    indices: CharIndices<'a>,
    tag_state: TagState,
    last_char: (usize, char),
}

struct Reader for MicroXml - the Class
Rust has Structs + Traits, but for me it is just like Class/Object.
Just without inheritance.
All the fields are internal and not public.
The only way to interact is through methods.

Fields
input: &'a str

reference to the xml string (no allocation)

indices: CharIndices<'a>

Iterator CharIndices over the input string

tag_state: TagState

I need to know the TagState for programming as a state machine

last_char: (usize, char)

the last read character from the iterator

Methods

impl<'a, '_> ReaderForMicroXml<'_>

pub fn new(input: &str) -> ReaderForMicroXml

Constructor. String is immutably borrowed here. No allocation.

pub fn read_event(&mut self) -> Event

Reads the next event: StartElement, Attribute, Text, EndElement
If the Option None (Eof) has propagated till here, this is an Error.

fn read_event_internal(&mut self) -> Option<Event>

Reads the next event (internal).
The internal function can understand when the Eof is in a correct position
and stops the propagation of Option None.

fn read_element_name(&mut self) -> Option<Event>

Reads the element name
Propagation of Option None if is Eof

fn read_attribute(&mut self) -> Option<Event>

Reads the attribute name and value.
Return Option None if Eof.

fn read_end_element(&mut self) -> Option<Event>

reads end element

fn read_text_node(&mut self, start_pos: usize) -> Option<Event>

Reads text node
I don't do any encoding/decoding here, because I need it "as is" for html templating.
I preserve all the "significant" whitespaces because I will use this for templating.
And because there is no hard standard for trailing spaces in xml text node.
If reached Eof propagates Option None.

fn skip_comment(&mut self) -> Option<usize>

Comments are not data for MicroXml standard,
They are ignored, I don't return them.
The Option is returned only because of Option None propagation because of Eof.

fn move_next_char(&mut self) -> Option<usize>

Moves the iterator and stores the last_char.
Iterator next() of CharIndices is consuming the char.
There is no way back to the same char.
But often I need to get again the same character of the last operation.
I tried with peekable.peek(), but it gives a reference and this was a problem.
So now I have 2 separate methods: move_next_char() and get_last_char().
I store the last_char for repeated use.
Anytime it can reach the End of File (Eof),
then it propagates the Option None to the caller with the ? syntax.
Only the caller knows if the Eof here is ok or it is an unexpected error.
The usize inside the Option is only a dummy,
only because I need to propagate the Option None because of Eof

fn get_last_char(&self) -> (usize, char)

Returns the last_char, but doesn't move the iterator.
Cannot error

fn skip_whitespace_and_get_last_char(&mut self) -> Option<(usize, char)>

Skips all whitespaces if there is any
and returns the last_char when it is not whitespace.
The caller must know prior to call that the whitespaces are insignificant.
If found Eof, propagates Option None.